+++ packages/pwa-kit-runtime/src/utils/ssr-proxying.js
@@ -723,7 +723,7 @@ export const rewriteProxyResponseHeaders = ({
  * @private
  * @type {string[]}
  */
-export const X_HEADERS_TO_REMOVE_PROXY = ['x-mobify-access-key']
+export const X_HEADERS_TO_REMOVE_PROXY = ['x-mobify-access-key', 'x-sfdc-access-control']
 
 /**
  * List of x- headers that are removed from origin requests.